fix: move prompt util functiont to prompt
prompt_and_delete was a function in utils which also made use of read_user_yes_no
which was in prompt.
This means that prompt could not import utils because it would create a circular import.
This change moves prompt_and_delete to prompt, so that prompt may make use of util functions
