+++ cookiecutter/prompt.py
@@ -1,12 +1,14 @@
 """Functions for prompting the user for project info."""
 import json
 from collections import OrderedDict
+import os
+import sys
 
 from rich.prompt import Prompt, Confirm, PromptBase, InvalidResponse
 from jinja2.exceptions import UndefinedError
 
 from cookiecutter.exceptions import UndefinedVariableInTemplate
-from cookiecutter.utils import create_env_with_context
+from cookiecutter.utils import create_env_with_context, rmtree
 
 
 def read_user_variable(var_name, default_value, prompts=None, prefix=""):
@@ -313,3 +315,41 @@ def prompt_for_config(context, no_input=False):
             raise UndefinedVariableInTemplate(msg, err, context) from err
 
     return cookiecutter_dict
+
+
+def prompt_and_delete(path, no_input=False):
+    """
+    Ask user if it's okay to delete the previously-downloaded file/directory.
+
+    If yes, delete it. If no, checks to see if the old version should be
+    reused. If yes, it's reused; otherwise, Cookiecutter exits.
+
+    :param path: Previously downloaded zipfile.
+    :param no_input: Suppress prompt to delete repo and just delete it.
+    :return: True if the content was deleted
+    """
+    # Suppress prompt if called via API
+    if no_input:
+        ok_to_delete = True
+    else:
+        question = (
+            f"You've downloaded {path} before. Is it okay to delete and re-download it?"
+        )
+
+        ok_to_delete = read_user_yes_no(question, 'yes')
+
+    if ok_to_delete:
+        if os.path.isdir(path):
+            rmtree(path)
+        else:
+            os.remove(path)
+        return True
+    else:
+        ok_to_reuse = read_user_yes_no(
+            "Do you want to re-use the existing version?", 'yes'
+        )
+
+        if ok_to_reuse:
+            return False
+
+        sys.exit()
