+++ cookiecutter/utils.py
@@ -4,7 +4,6 @@ import logging
 import os
 import shutil
 import stat
-import sys
 import tempfile
 from pathlib import Path
 from typing import Dict
@@ -12,7 +11,6 @@ from typing import Dict
 from jinja2.ext import Extension
 from cookiecutter.environment import StrictEnvironment
 
-from cookiecutter.prompt import read_user_yes_no
 
 logger = logging.getLogger(__name__)
 
@@ -69,44 +67,6 @@ def make_executable(script_path):
     """
     status = os.stat(script_path)
     os.chmod(script_path, status.st_mode | stat.S_IEXEC)
-
-
-def prompt_and_delete(path, no_input=False):
-    """
-    Ask user if it's okay to delete the previously-downloaded file/directory.
-
-    If yes, delete it. If no, checks to see if the old version should be
-    reused. If yes, it's reused; otherwise, Cookiecutter exits.
-
-    :param path: Previously downloaded zipfile.
-    :param no_input: Suppress prompt to delete repo and just delete it.
-    :return: True if the content was deleted
-    """
-    # Suppress prompt if called via API
-    if no_input:
-        ok_to_delete = True
-    else:
-        question = (
-            f"You've downloaded {path} before. Is it okay to delete and re-download it?"
-        )
-
-        ok_to_delete = read_user_yes_no(question, 'yes')
-
-    if ok_to_delete:
-        if os.path.isdir(path):
-            rmtree(path)
-        else:
-            os.remove(path)
-        return True
-    else:
-        ok_to_reuse = read_user_yes_no(
-            "Do you want to re-use the existing version?", 'yes'
-        )
-
-        if ok_to_reuse:
-            return False
-
-        sys.exit()
 
 
 def simple_filter(filter_function):
